refactor: merge global-general into root config
