+++ cyberdrop_dl/crawlers/onefichier.py
@@ -106,7 +106,7 @@ class OneFichierCrawler(Crawler):
 
     async def _request_download(self, url: AbsoluteHttpURL, password: str | None) -> AbsoluteHttpURL:
         data = {"pass": password} if password else {}
-        if not self.config.global_settings.general.ssl_context:
+        if not self.config.ssl_context:
             data["dl_no_ssl"] = "on"
 
         soup = await self.request_soup(url, method="POST", data=data or None)
